+++ pkg/vault/contracts/test/PoolFactoryMock.sol
@@ -10,14 +10,13 @@ import { IRateProvider } from "@balancer-labs/v3-interfaces/contracts/vault/IRat
 import "@balancer-labs/v3-interfaces/contracts/vault/VaultTypes.sol";
 
 import { FactoryWidePauseWindow } from "../factories/FactoryWidePauseWindow.sol";
-import { PoolConfigBits } from "../lib/PoolConfigLib.sol";
 
 contract PoolFactoryMock is FactoryWidePauseWindow {
     uint256 private constant DEFAULT_SWAP_FEE = 0;
 
     IVault private immutable _vault;
 
-    constructor(IVault vault, uint256 pauseWindowDuration) FactoryWidePauseWindow(pauseWindowDuration) {
+    constructor(IVault vault, uint32 pauseWindowDuration) FactoryWidePauseWindow(pauseWindowDuration) {
         _vault = vault;
     }
 
@@ -46,7 +45,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
         address pool,
         TokenConfig[] memory tokenConfig,
         uint256 swapFee,
-        uint256 pauseWindowDuration,
+        uint32 pauseWindowDuration,
         PoolRoleAccounts memory roleAccounts,
         address poolHooksContract
     ) external {
@@ -54,7 +53,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
             pool,
             tokenConfig,
             swapFee,
-            block.timestamp + pauseWindowDuration,
+            uint32(block.timestamp) + pauseWindowDuration,
             roleAccounts,
             poolHooksContract,
             LiquidityManagement({
@@ -105,7 +104,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
     function registerPoolAtTimestamp(
         address pool,
         TokenConfig[] memory tokenConfig,
-        uint256 timestamp,
+        uint32 timestamp,
         PoolRoleAccounts memory roleAccounts,
         address poolHooksContract,
         LiquidityManagement calldata liquidityManagement
