+++ pkg/vault/test/foundry/RecoveryMode.t.sol
@@ -148,7 +148,7 @@ contract RecoveryModeTest is BaseVaultTest {
         assertFalse(vault.isPoolInRecoveryMode(pool), "Pool should not be in Recovery Mode after pausing");
 
         // Ensure we are in the permissionless period of the Pool.
-        (, , uint256 bufferPeriodEndTime, ) = vault.getPoolPausedState(pool);
+        (, , uint32 bufferPeriodEndTime, ) = vault.getPoolPausedState(pool);
 
         skip(bufferPeriodEndTime);
